Set EXTENT value to full epsg:3857 bounds
Since the default spatial reference for Imposm and Basemaps is
EPSG:3857 this change sets the EXTENT value to the full EPSG:3544
World Area of Use bounds which cover unprojected coordinates
`-180 -85.05112878 180 85.05112878` (see https://www.epsg-registry.org).

The difference is small, the current settings are equivalent to about
`-179.66305682 -85.02197648 179.66305682 85.02197648`!

This is for the East of Fiji and Russia.